homework2/tasks/hw5.py
"""
Some of the functions have a bit cumbersome behavior when we deal with
positional and keyword arguments.

Write a function that accept any iterable of unique values and then
it behaves as range function:


import string


assert = custom_range(string.ascii_lowercase, 'g') == ['a', 'b', 'c', 'd', 'e', 'f']
assert = custom_range(string.ascii_lowercase, 'g', 'p') == ['g', 'h', 'i', 'j', 'k', 'l', 'm', 'n', 'o']
assert = custom_range(string.ascii_lowercase, 'p', 'g', -2) == ['p', 'n', 'l', 'j', 'h']

"""
from typing import List


def custom_range(iterable, start=None, end=None, step=1) -> List:
    if not start:
        start = iterable[0]
    start_index = iterable.index(start)
    end_index = iterable.index(end)
    return list(iterable[start_index:end_index:step])
